refactor(desktop): make an unscreened open path unrepresentable
The desktop shell screens every path the OS hands it — the NTFS
alternate-data-stream scan, UNC refusal, the extension allowlist and the
regular-file check all live in `validate_open_candidate` — and then
throws the result away: the screener returned a bare `PathBuf`, so
`PendingOpens`, `promote_healthy_and_drain`, `try_queue_or_post`,
`post_drained_paths` and `cold_start_file` all carried a type that says
nothing about where it came from. Nothing distinguished "this passed the
screener" from "this came from somewhere", and the guarantee rested on
call-site discipline plus a doc comment — the same shape of guarantee
#1344 already found insufficient once, when the macOS `RunEvent::Opened`
surface ran neither check for months.

`ScreenedOpenPath` is a newtype whose tuple field is private, so
`validate_open_candidate` is the only thing in the program that can mint
one. The module boundary is the mechanism, not tidying: `lib.rs` is
~6,900 lines in a single module with eight `#[cfg(test)]` submodules,
and a newtype declared there would be constructible on every one of
those lines — `PendingOpens`' own "private" field is written straight
through by `lib.rs`'s tests today. So the whole open-candidate cluster
moves to `src-tauri/src/open_candidate.rs`, and the newtype's privacy is
enforced by rustc rather than by review.

`Deref<Target = Path>` is taken deliberately, against the usual newtype
guidance, so the macOS-gated call sites this box cannot type-check
change by zero lines.

The screener stays `pub(crate)` (its unit tests are a sibling module) and
now states its precondition: `debug_assert!` that the argument is a
resolved path, because the ADS scan's drive-letter exemption at byte
index 1 is unsound on a relative candidate such as `f:ADS.md`.

Closes #1415
